Update global-ci workflows to also be able to test UI run-ui-tests
Since the global-ci will run for any PR that changes the run-ui-tests
action in the tackle2-ui repository, we want to be able to use the
those changes in the global-ci run.  This is done by checking out the
tackle2-ui repository at a specific ref and running the run-ui-tests
action "locally".

See CI fails in https://github.com/konveyor/tackle2-ui/pull/2787
for the motivation for this change.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>